# docs/swagger.yaml
openapi: 3.0.3
info:
  title: AlpineLink 雪域 API
  description: |
    AlpineLink 雪域高端登山户外社交平台后端 API 文档。

    ## 认证方式
    大多数接口需要在请求头中携带 JWT Token：
    ```
    Authorization: Bearer <token>
    ```
    Token 通过 `/api/auth/login` 获取。

    ## 接口分级
    - **P0**：核心功能，认证模块
    - **P1**：主要业务，帖子、评论、山峰/路线、活动、消息
    - **P2**：扩展功能，向导、俱乐部、SOS、后台统计

    ## 统一响应格式
    ```json
    { "code": 0, "message": "success", "data": {} }
    ```

    ## WebSocket 实时消息
    连接地址：`ws://<host>/ws?token=<jwt_token>`

    消息格式：
    ```json
    { "type": "message", "chatId": 1, "content": "你好", "msgType": "text" }
    ```

    服务端推送格式：
    ```json
    { "type": "message", "data": { "id": 1, "chatId": 1, "senderId": 2, "content": "你好", "msgType": "text", "createdAt": "2026-01-01T00:00:00Z" } }
    ```
  version: 1.0.0
  contact:
    name: AlpineLink CEO
    url: https://desktop-tutorial-production-182a.up.railway.app
servers:
  - url: https://desktop-tutorial-production-182a.up.railway.app
    description: Railway 生产环境
  - url: http://localhost:3000
    description: 本地开发环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token，通过 `/api/auth/login` 获取
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 1003
        message:
          type: string
          example: 未登录或Token无效
        data:
          nullable: true
          example: null
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
          example: 1
        otherId:
          type: integer
          example: 2
        otherName:
          type: string
          example: 张三
        otherAvatar:
          type: string
          format: uri
          nullable: true
          example: https://example.com/avatar.jpg
        lastMsg:
          type: string
          nullable: true
          example: 明天几点出发？
        unread:
          type: integer
          example: 3
        updated_at:
          type: string
          format: date-time
          example: "2026-01-20T14:00:00Z"
    PrivateMessage:
      type: object
      properties:
        id:
          type: integer
          example: 1
        sender_id:
          type: integer
          example: 1
        content:
          type: string
          example: 明天几点出发？
        type:
          type: string
          enum: [text, image, mixed]
          example: text
        images:
          type: array
          items:
            type: string
            format: uri
          example: []
        is_read:
          type: integer
          enum: [0, 1]
          example: 0
        created_at:
          type: string
          format: date-time
          example: "2026-01-20T14:00:00Z"

    # ────────── 通知 ──────────
...
  responses:
    Unauthorized:
      description: 未登录或 Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1003
            message: 未登录或Token无效
            data: null
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1004
            message: 权限不足
            data: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1005
            message: 资源不存在
            data: null
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 1001
            message: 参数缺失
            data: null

# ══════════════════════════════════════════════════════
# PATHS
# ══════════════════════════════════════════════════════
paths:
...
  /api/messages/conversations:
    get:
      tags: [Messages]
      summary: 获取当前用户所有会话列表（需登录）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 会话列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Messages]
      summary: 创建或获取与指定用户的会话（需登录）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_user_id]
              properties:
                target_user_id:
                  type: integer
                  example: 2
      responses:
        '200':
          description: 返回会话对象
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/messages/conversations/{id}/messages:
    get:
      tags: [Messages]
      summary: 获取会话消息列表（需登录）
      description: 返回会话内所有消息，并将对方消息标记为已读。
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 会话ID
          schema:
            type: integer
      responses:
        '200':
          description: 消息列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrivateMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Messages]
      summary: 发送私信（需登录）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 会话ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: 明天几点出发？
                type:
                  type: string
                  enum: [text, image, mixed]
                  example: text
                images:
                  type: array
                  items:
                    type: string
                    format: uri
                  example: []
      responses:
        '200':
          description: 发送成功，返回消息对象
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/messages/conversations/{id}/messages/poll:
    get:
      tags: [Messages]
      summary: 长轮询获取新消息（需登录）
      description: 返回 id > after 的新消息，用于轻量级实时刷新。
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: after
          in: query
          description: 上一条已知消息的 id，返回比它更新的消息
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 新消息列表（可能为空数组）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrivateMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ────────── 5.5 通知模块 ──────────
...